chore(ugc): ✨ add confirmation modal before media deletion
A confirmation modal is now presented to the user before deleting media. The modal asks the user to confirm the deletion or cancel the operation. If the user confirms (chooses the "Elimina" option), the media is deleted. If the user cancels, the deletion is aborted.

This change enhances user interaction by preventing accidental deletions and providing a clear option to cancel the action.

+++ projects/wm-core/src/store/features/ugc/ugc.effects.ts
@@ -320,15 +320,38 @@ export class UgcEffects {
       ofType(deleteUgcMedia),
       mergeMap(({media}) => {
         if(media?.id) {
-          return from(this._ugcSvc.deleteApiMedia(media.id)).pipe(
-            mergeMap(() => [
-              deleteUgcMediaSuccess({media}),
-              syncUgc(),
-            ]),
-            catchError(error => of(deleteUgcMediaFailure({error: error.error.message}))),
-          )
+          return from(this._alertCtrl.create({
+            header: this._langSvc.instant('Conferma eliminazione'),
+            message: this._langSvc.instant('Sei sicuro di voler eliminare questo media?'),
+            buttons: [
+              {
+                text: this._langSvc.instant('Annulla'),
+                role: 'cancel',
+              },
+              {
+                text: this._langSvc.instant('Elimina'),
+                role: 'destructive',
+              },
+            ],
+          })).pipe(
+            switchMap(modal => from(modal.present()).pipe(
+              switchMap(() => from(modal.onDidDismiss()))
+            )),
+            switchMap((result) => {
+              if (result.role === 'destructive') {
+                return from(this._ugcSvc.deleteApiMedia(media.id)).pipe(
+                  mergeMap(() => [
+                    deleteUgcMediaSuccess({media}),
+                    syncUgc(),
+                  ]),
+                  catchError(error => of(deleteUgcMediaFailure({error: error.error.message}))),
+                );
+              }
+              return EMPTY;
+            }),
+          );
         }
-        return of(deleteUgcMediaFailure({error: 'Media ID not found'}))
+        return of(deleteUgcMediaFailure({error: 'Media ID not found'}));
       }),
       catchError(error => of(deleteUgcMediaFailure({error}))),
     ),
